+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `--verbose`/`-v` flag (count-based): enables verbose debug output
-- `--debug` enables debug logging
+- `--debug`/`-vv` enables debug logging
 
 ## [0.1.4] - 2026-03-27